use a 16bit index for tile graphics (hope I didn't break anything)

+++ source/skyland/tileId.hpp
@@ -32,7 +32,7 @@ namespace skyland
 
 
 
-using TileId = u8;
+using TileId = u16;
